+++ frontend/src/components/editing-tools/maskSelector.js
@@ -8,7 +8,7 @@ export default class MaskSelector {
     this.redoStore = useRedoStore();
     this.fileName = fileName;
   }
-  async removeObject(id, undo = false, image, mask, transparentBackground, basePixels = null) {
+  async removeObject(id, undo = false, image, mask, transparentBackground, restoreBackGround, basePixels = null) {
     if (undo) {
       this.redoStore.addFileFuture(this.fileName, {id:[id], action:'remove', undo:true});
     } else {
@@ -22,7 +22,7 @@ export default class MaskSelector {
       const g = editSetup.maskData[i + 1];
       const b = editSetup.maskData[i + 2];
       if (r > 0 && g > 0 && b > 0) {
-        if (undo && basePixels) {
+        if (undo || restoreBackGround && basePixels) {
           pixelRemovingTools.restorePixel(editSetup.shownPixels, basePixels, i);
         } else {
           transparentBackground ? 
@@ -38,7 +38,7 @@ export default class MaskSelector {
     return editSetup.canvas.toDataURL('image/png').split(',')[1];
   }
 
-  async selectOnlyObject(id, undo = false, image, mask, transparentBackground, basePixels = null) {
+  async selectOnlyObject(id, undo = false, image, mask, transparentBackground, restoreBackground, basePixels = null) {
     if (undo) {
       this.redoStore.addFileFuture(this.fileName, {id:[id], action:'select', undo:true});
     } else {
@@ -52,7 +52,7 @@ export default class MaskSelector {
       const b = editSetup.maskData[i + 2];
 
       if (r == 0 && g == 0 && b == 0) {
-        if (undo && basePixels) {
+        if (undo || restoreBackground && basePixels) {
           pixelRemovingTools.restorePixel(editSetup.shownPixels, basePixels, i);
         } else {
           transparentBackground ?